requirements.txt
firecrawl-py
openai
python-dotenv
pandas
openpyxl
requests
beautifulsoup4
tiktoken
